requirements_transcricao.txt
# ============================================
# Projeto: Tele_IA Transcrição (Pipeline v2)
# Ambiente estável e compatível (Python 3.12)
# ============================================

whisperx==3.7.2
faster-whisper==0.9.0
pyannote.audio==3.1.1
torch==2.2.0
torchaudio==2.2.0
ctranslate2==4.6.0
onnxruntime==1.23.1
speechbrain==1.0.3
deepmultilingualpunctuation==1.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
jiwer==4.0.0
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.1.1
matplotlib==3.9.2
